requirements.txt
pypyodbc
pyserial